+++ rocon_app_utilities/src/rocon_app_utilities/rapp.py
@@ -9,103 +9,12 @@ from __future__ import division, print_function
 import yaml
 from .exceptions import *
 import rocon_uri
+from .rapp_validation import classify_rapp_type 
+from .rapp_loader import load_rapp_yaml_from_file
 
 #################################################################################
 # Local Method
 #################################################################################
-
-
-def _is_implementation_rapp(data):
-    '''
-      It is implementation if it contains compatibility and launch attributes
-
-      :param data: dict of rapp definition
-      :type data: dict
-
-      :returns: whether it is implementation or not
-      :rtype: bool
-    '''
-    IMPLEMETATION_VALIDATION_LIST = ['launch', 'compatibility']
-
-    r = set(IMPLEMETATION_VALIDATION_LIST)
-    m = set(data.keys())
-
-    return r.issubset(m)
-
-
-def _is_ancestor_rapp(data):
-    '''
-      It is ancestor rapp if it does not have parent_name attribute
-
-      :param data: dict of rapp definition
-      :type data: dict
-
-      :returns: whether it is ancestor or not
-      :rtype: bool
-    '''
-    CHILD_VALIDATION_LIST = ['parent_name']
-
-    r = set(CHILD_VALIDATION_LIST)
-    m = set(data.keys())
-    return (not r.issubset(m))
-
-
-def load_yaml_from_file(filename):
-    '''
-      Load rapp specs yaml from the given file
-
-      :param filename: absolute path to the rapp definition
-      :type filename: str
-
-      :returns: dict of loaded rapp
-      :rtype: dict
-
-      :raises: InvalidRappFieldException: Rapp includes invalid filed
-    '''
-    RAPP_ATTRIBUTES = ['display', 'description', 'icon', 'public_interface', 'public_parameters', 'compatibility', 'launch', 'parent_name', 'pairing_clients', 'required_capability']
-
-    with open(filename, 'r') as f:
-        app_data = yaml.load(f.read())
-
-        for d in app_data:
-            if d not in RAPP_ATTRIBUTES:
-                raise InvalidRappException('Invalid Field : [' + str(d) + '] Valid Fields : [' + str(RAPP_ATTRIBUTES) + ']')
-
-    return app_data
-
-
-def classify_rapp_type(data):
-    '''
-      Classify the current rapp among VirtualAncestor, ImplementationAnacestor, ImplementationChild
-
-      :param data: rapp yaml data
-      :type data: dict
-
-      :returns: its classification is implementation? is ancestor? and its type
-      :rtypes: bool, bool, str
-
-      :raises: InvalidRappException: It is virtual child rapp
-      :raises: InvalidRappFieldException: Rapp contains invalid field
-    '''
-    is_impl = _is_implementation_rapp(data)
-    is_ance = _is_ancestor_rapp(data)
-
-    impl = 'Implementation' if is_impl else 'Virtual'
-    ance = 'Ancestor' if is_ance else 'Child'
-    try:
-        if is_impl and is_ance:  # Implementation Ancestor
-            ImplementationAncestorRapp.is_valid(data)
-        elif is_impl and not is_ance:  # Implementation Child
-            ImplementationChildRapp.is_valid(data)
-        elif not is_impl and is_ance:  # Virtual Ancestor
-            VirtualAncestorRapp.is_valid(data)
-        else:                         # Virtual Child
-            raise InvalidRappException('Virtual Child rapp. Invalid!')
-    except InvalidRappFieldException as ife:
-        raise ife
-
-    t = str(impl + ' ' + ance)
-    return is_impl, is_ance, t
 
 
 def validate_rapp_field(data):
@@ -117,14 +26,20 @@ def validate_rapp_field(data):
     pass
 
 
+#################################################################################
+# Rapp Class 
+#################################################################################
+
+
 class Rapp(object):
     '''
         Rocon(or Robot) App definition.
     '''
-    __slots__ = ['resource_name', 'data', 'type', 'is_implementation', 'is_ancestor', 'ancestor_name', 'parent_name']
+    __slots__ = ['resource_name', 'raw_data', 'data', 'type', 'is_implementation', 'is_ancestor', 'ancestor_name', 'parent_name']
 
     def __init__(self, name, filename=None):
         self.resource_name = name
+        self.raw_data = {}
         self.data = {}
         self.type = None
         self.ancestor_name = None
@@ -132,7 +47,7 @@ class Rapp(object):
         self.is_ancestor = False
 
         if filename:
-            self.load_from_file(filename)
+            self.load_spec_from_file(filename)
             #validate_rapp_field(self.dat)
             self.classify()
 
@@ -150,7 +65,7 @@ class Rapp(object):
           :returns: true if compatible
           :rtype: bool
         '''
-        my_uri = self.data['compatibility'] if 'compatibility' in self.data else None
+        my_uri = self.raw_data['compatibility'] if 'compatibility' in self.raw_data else None
 
         if my_uri:
             return rocon_uri.is_compatible(my_uri, uri)
@@ -166,17 +81,17 @@ class Rapp(object):
           is_ancestor        : Is it Ancestor?
           type               : str(Virtual Ancestor, Implementation Ancestor or Implementation Child)
         '''
-        self.parent_name = self.data['parent_name'] if 'parent_name' in self.data else None
-        self.is_implementation, self.is_ancestor, self.type = classify_rapp_type(self.data)
+        self.parent_name = self.raw_data['parent_name'] if 'parent_name' in self.raw_data else None
+        self.is_implementation, self.is_ancestor, self.type = classify_rapp_type(self.raw_data)
 
-    def load_from_file(self, filename):
+    def load_spec_from_file(self, filename):
         '''
           loads rapp data from file. and classifies itself.
 
           :param filename: absolute path to rapp definition
           :type filename: str
         '''
-        self.data = load_yaml_from_file(filename)
+        self.raw_data = load_rapp_yaml_from_file(filename)
         self.classify()
 
     def inherit(self, rapp):
@@ -189,65 +104,10 @@ class Rapp(object):
         INHERITABLE_ATTRIBUTES = ['display', 'description', 'icon', 'public_interface', 'public_parameters', 'parent_name']
 
         # Once it inherits, it removes parent_specification field. If it is inherits from another child, it obtains parent_specification anyway
-        del self.data['parent_name']
+        del self.raw_data['parent_name']
 
         for attribute in INHERITABLE_ATTRIBUTES:
-            if not attribute in self.data and attribute in rapp.data:
-                self.data[attribute] = rapp.data[attribute]
+            if not attribute in self.raw_data and attribute in rapp.raw_data:
+                self.raw_data[attribute] = rapp.raw_data[attribute]
 
         self.classify()
-
-
-class RappValidation(Rapp):
-    _required = []
-    _optional = []
-    _not_allowed = []
-
-    @classmethod
-    def is_valid(cls, data):
-        '''
-            Rapp Validation. If it has all requirements and does not include any not_allowed attributes, it is valid rapp
-
-            :param data: rapp specification
-            :type data: dict
-
-            :returns: whether it is valid spec
-            :rtype: bool
-
-            :raises: InvalidRappFieldException: It includes not allowed field or misses required field
-        '''
-        missing_required = cls._difference(cls._required, data.keys())
-        included_not_allowed = cls._intersection(cls._not_allowed, data.keys())
-
-        if len(missing_required) > 0 or len(included_not_allowed) > 0:
-            raise InvalidRappFieldException(cls, missing_required, included_not_allowed)
-
-        return True
-
-    @classmethod
-    def _intersection(cls, attributes, data):
-        intersection = set(attributes).intersection(set(data))
-        return list(intersection)
-
-    @classmethod
-    def _difference(cls, attributes, data):
-        diff = set(attributes).difference(set(data))
-        return list(diff)
-
-
-class VirtualAncestorRapp(RappValidation):
-    _required = ['display', 'description']
-    _optional = ['icon', 'public_interface', 'public_parameters']
-    _not_allowed = ['compatibility', 'launch', 'parent_name', 'pairing_clients', 'required_capability']
-
-
-class ImplementationAncestorRapp(RappValidation):
-    _required = ['display', 'description', 'compatibility', 'launch']
-    _optional = ['icon', 'pairing_clients', 'required_capability', 'public_interface', 'public_parameters']
-    _not_allowed = ['parent_name']
-
-
-class ImplementationChildRapp(RappValidation):
-    _required = ['compatibility', 'launch', 'parent_name']
-    _optional = ['icon', 'pairing_clients', 'required_capability']
-    _not_allowed = ['public_interface', 'public_parameters']
